64 line, multiple display, maybe malloc
Several changes here.   Still based on SPIN library...

supports multiple displays.  Needed to move screen buffer into object.

Also optionally supports 64 line displays.  Did this as my only 2nd of
these type of display has 64 lines...

Did not want to have to Edit library to support either 32 or 64 line
displays like the Adafruit library, so takes this as a parameter.  So if
64 line display option is used all displays allocate enough memory for
64 lines even if only 32 bit displays... So made optional a way to use
malloc instead of fixed buffer, so it only allocated the memory it
needs.